requirements.txt
fastapi==0.115.6
uvicorn==0.22.0
python-dotenv==1.0.0
notion-client==2.2.0
pydantic==2.10.0
pydantic-ai==0.0.18
httpx>=0.27.2
python-multipart==0.0.6